move customerPu and embeddable tests
