--- jpa/eclipselink.jpa.test/src/it/java/org/eclipse/persistence/testing/models/jpa/advanced/embeddable/Description.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0,
- * or the Eclipse Distribution License v. 1.0 which is available at
- * http://www.eclipse.org/org/documents/edl-v10.php.
- *
- * SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
- */
-
-// Contributors:
-//     03/19/2018-2.7.2 Lukas Jungmann
-//       - 413120: Nested Embeddable Null pointer
-//       - 496836: NullPointerException on ObjectChangeSet.mergeObjectChanges
-package org.eclipse.persistence.testing.models.jpa.advanced.embeddable;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-
-@Embeddable
-public class Description {
-
-    private String note;
-
-    public Description() {
-    }
-    
-    public Description(String note) {
-        this.note = note;
-    }
-
-    @Column(name = "NOTE")
-    public String getNote() {
-        return note;
-    }
-
-    public void setNote(String note) {
-        this.note = note;
-    }
-    
-    
-}
