--- jpa/eclipselink.jpa.test/src/it/resources/eclipselink-annotation-model/orm-invalid-named-query.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2018, 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0,
-    or the Eclipse Distribution License v. 1.0 which is available at
-    http://www.eclipse.org/org/documents/edl-v10.php.
-
-    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
-
--->
-
-<entity-mappings version="1.0"
-    xmlns="http://java.sun.com/xml/ns/persistence/orm" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://java.sun.com/xml/ns/persistence/orm orm_1_0.xsd">
-    <description>Mapping Document containing persistence unit metadata.
-        Do not provide any class specific mapping information here.
-    </description>
-    <named-query name="non-existant-entity">
-        <query>SELECT a FROM Alien a</query>
-    </named-query>
-</entity-mappings>
