+++ jpa/eclipselink.jpa.testapps/jpa.test.advanced.embeddable/src/main/java/org/eclipse/persistence/testing/models/jpa/advanced/embeddable/Visitor.java
