+++ jpa/eclipselink.jpa.testapps/jpa.test.advanced.embeddable/src/main/resources-ejb/META-INF/ejb-jar.xml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2019, 2022 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0 which is available at
+    http://www.eclipse.org/legal/epl-2.0,
+    or the Eclipse Distribution License v. 1.0 which is available at
+    http://www.eclipse.org/org/documents/edl-v10.php.
+
+    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+
+-->
+
+<ejb-jar version="4.0" xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee https://jakarta.ee/xml/ns/jakartaee/ejb-jar_4_0.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="https://jakarta.ee/xml/ns/jakartaee">
+    <enterprise-beans>
+        <session>
+            <ejb-name>GenericTestRunner</ejb-name>
+            <persistence-context-ref>
+                <persistence-context-ref-name>persistence/embeddable/entity-manager</persistence-context-ref-name>
+                <persistence-unit-name>embeddable</persistence-unit-name>
+            </persistence-context-ref>
+            <persistence-context-ref>
+                <persistence-context-ref-name>persistence/invalid-named-query/entity-manager</persistence-context-ref-name>
+                <persistence-unit-name>invalid-named-query</persistence-unit-name>
+            </persistence-context-ref>
+            <persistence-context-ref>
+                <persistence-context-ref-name>persistence/invalid-named-query-tolerate/entity-manager</persistence-context-ref-name>
+                <persistence-unit-name>invalid-named-query-tolerate</persistence-unit-name>
+            </persistence-context-ref>
+            <persistence-unit-ref>
+                <persistence-unit-ref-name>persistence/embeddable/factory</persistence-unit-ref-name>
+                <persistence-unit-name>embeddable</persistence-unit-name>
+            </persistence-unit-ref>
+            <persistence-unit-ref>
+                <persistence-unit-ref-name>persistence/invalid-named-query/factory</persistence-unit-ref-name>
+                <persistence-unit-name>invalid-named-query</persistence-unit-name>
+            </persistence-unit-ref>
+            <persistence-unit-ref>
+                <persistence-unit-ref-name>persistence/invalid-named-query-tolerate/factory</persistence-unit-ref-name>
+                <persistence-unit-name>invalid-named-query-tolerate</persistence-unit-name>
+            </persistence-unit-ref>
+        </session>
+    </enterprise-beans>
+</ejb-jar>
