+++ jpa/eclipselink.jpa.testapps/jpa.test.advanced.embeddable/src/main/resources/META-INF/persistence.xml
@@ -1,0 +1,51 @@
+<!--
+
+    Copyright (c) 2019, 2022 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0 which is available at
+    http://www.eclipse.org/legal/epl-2.0,
+    or the Eclipse Distribution License v. 1.0 which is available at
+    http://www.eclipse.org/org/documents/edl-v10.php.
+
+    SPDX-License-Identifier: EPL-2.0 OR BSD-3-Clause
+
+-->
+
+<persistence xmlns="http://java.sun.com/xml/ns/persistence" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://java.sun.com/xml/ns/persistence persistence_1_0.xsd" version="1.0">
+
+    <persistence-unit name="embeddable">
+        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
+        <class>org.eclipse.persistence.testing.models.jpa.advanced.embeddable.Visitor</class>
+        <exclude-unlisted-classes>true</exclude-unlisted-classes>
+    </persistence-unit>
+
+    <persistence-unit name="invalid-named-query">
+        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
+        <mapping-file>META-INF/orm-invalid-named-query.xml</mapping-file>
+        <class>org.eclipse.persistence.testing.models.jpa.advanced.embeddable.Visitor</class>
+        <exclude-unlisted-classes>true</exclude-unlisted-classes>
+    </persistence-unit>
+
+    <persistence-unit name="invalid-named-query-tolerate">
+        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
+        <mapping-file>META-INF/orm-invalid-named-query.xml</mapping-file>
+        <class>org.eclipse.persistence.testing.models.jpa.advanced.embeddable.Visitor</class>
+        <exclude-unlisted-classes>true</exclude-unlisted-classes>
+        <properties>
+            <property name="eclipselink.tolerate-invalid-jpql" value="true"/>
+        </properties>
+    </persistence-unit>
+
+    <persistence-unit name="invalid-named-query-validation-only">
+        <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
+        <mapping-file>META-INF/orm-invalid-named-query.xml</mapping-file>
+        <class>org.eclipse.persistence.testing.models.jpa.advanced.embeddable.Visitor</class>
+        <exclude-unlisted-classes>true</exclude-unlisted-classes>
+        <properties>
+            <property name="eclipselink.deploy-on-startup" value="false"/>
+            <property name="eclipselink.tolerate-invalid-jpql" value="true"/>
+            <property name="eclipselink.validation-only" value="true"/>
+        </properties>
+    </persistence-unit>
+</persistence>
